+++ cetmix_tower_server_queue/readme/HISTORY.md
@@ -1,3 +1,9 @@
+## 17.0.1.1.0 (2025-07-30)
+
+- Features: cetmix_tower_server_queue: Add async file upload/download via job queue (3720)
+- Features: Terminate command with error if job has failed (4718)
+
+
 ## 17.0.1.0.1 (2025-05-20)
 
 - Features: 'sudo' parameter is not passed to command. (4678)